update: ロボットサイズを更新

+++ robocon19_line_sensor/src/robocon19_line_sensor_node.cpp
@@ -11,14 +11,14 @@ enum
 
 int use_color = blue;
 
-int line_sensor_th[7] = {
-    0.6,
-    0.6,
-    0.6,
-    0.6,
-    0.6,
-    0.6,
-    0.6,
+float line_sensor_th[7] = {
+    0.6f,
+    0.6f,
+    0.6f,
+    0.6f,
+    0.6f,
+    0.6f,
+    0.6f,
 };
 
 // センサのデータを2値化
